+++ src/ui/junk-frame.lua
@@ -107,9 +107,10 @@ JunkFrame.frame = (function()
     labelText = L.DESTROY_NEXT_ITEM,
     onClick = Commands.destroy,
     onUpdateTooltip = function(self, tooltip)
-      local items = self:GetParent().items
-      if items and items[1] then
-        tooltip:SetBagItem(items[1].bag, items[1].slot)
+      tooltip:SetOwner(self, "ANCHOR_RIGHT")
+      local item = JunkFilter:GetNextDestroyableJunkItem()
+      if item then
+        tooltip:SetBagItem(item.bag, item.slot)
       end
     end
   })
@@ -168,21 +169,24 @@ JunkFrame.frame = (function()
       tooltip:AddLine(" ")
       tooltip:AddDoubleLine(L.LEFT_CLICK, L.SELL)
       tooltip:AddDoubleLine(L.RIGHT_CLICK, L.ADD_TO_LIST:format(Lists.PerCharExclusions.name))
-      tooltip:AddDoubleLine(Addon:Concat("+", L.SHIFT_KEY, L.LEFT_CLICK), Colors.Red(L.DESTROY))
       tooltip:AddDoubleLine(
         Addon:Concat("+", L.SHIFT_KEY, L.RIGHT_CLICK),
         L.ADD_TO_LIST:format(Lists.GlobalExclusions.name)
       )
+      tooltip:AddDoubleLine(Addon:Concat("+", L.ALT_KEY, L.RIGHT_CLICK), Colors.Red(L.DESTROY))
     end,
     itemButtonOnClick = function(self, button)
       if button == "LeftButton" then
-        local handler = IsShiftKeyDown() and Destroyer or Seller
-        handler:HandleItem(self.item)
+        Seller:HandleItem(self.item)
       end
 
       if button == "RightButton" then
-        local list = IsShiftKeyDown() and Lists.GlobalExclusions or Lists.PerCharExclusions
-        list:Add(self.item.id)
+        if IsAltKeyDown() then
+          Destroyer:HandleItem(self.item)
+        else
+          local list = IsShiftKeyDown() and Lists.GlobalExclusions or Lists.PerCharExclusions
+          list:Add(self.item.id)
+        end
       end
     end,
     getItems = function() return frame.items end,