+++ src/ui/merchant-button.lua
@@ -7,12 +7,9 @@ local JunkFilter = Addon:GetModule("JunkFilter")
 local L = Addon:GetModule("Locale")
 local StateManager = Addon:GetModule("StateManager")
 local TickerManager = Addon:GetModule("TickerManager")
-local Tooltip = Addon:GetModule("Tooltip")
 local Widgets = Addon:GetModule("Widgets")
 
-local LABEL_TEXT_FORMAT = ("%s %s"):format(ADDON_NAME, Colors.Grey("(%s)"):format(Colors.White("%s")))
-
-local junkItems = {}
+local LABEL_TEXT_FORMAT = Colors.Grey("(%s/%s)"):format(Colors.White("%s"), Colors.Red("%s"))
 
 -- ============================================================================
 -- Initialize
@@ -21,11 +18,33 @@ local junkItems = {}
 --- @class MerchantButtonWidget : ButtonWidget
 local frame = Widgets:Button({
   name = ADDON_NAME .. "_MerchantButton",
-  width = 100,
-  labelText = LABEL_TEXT_FORMAT:format(0),
+  width = 108,
+  labelText = LABEL_TEXT_FORMAT:format(0, 0),
   labelColor = Colors.Blue,
   enableClickHandling = true,
-  enableDragging = true
+  enableDragging = true,
+  onUpdateTooltip = function(this, tooltip)
+    tooltip:SetOwner(this, "ANCHOR_RIGHT")
+
+    if IsAltKeyDown() then
+      local item = JunkFilter:GetNextDestroyableJunkItem()
+      if item then
+        tooltip:SetBagItem(item.bag, item.slot)
+        tooltip:AddLine(" ")
+        tooltip:AddDoubleLine(L.RIGHT_CLICK, Colors.Red(L.DESTROY))
+        tooltip:Show()
+        return
+      end
+    end
+
+    tooltip:AddDoubleLine(Colors.Blue(ADDON_NAME), Colors.Grey(Addon.VERSION))
+    tooltip:AddLine(Addon:SubjectDescription(L.LEFT_CLICK, L.START_SELLING))
+    tooltip:AddLine(Addon:SubjectDescription(L.RIGHT_CLICK, L.TOGGLE_OPTIONS_FRAME))
+    tooltip:AddLine(Addon:SubjectDescription(Addon:Concat("+", L.SHIFT_KEY, L.LEFT_CLICK), L.TOGGLE_JUNK_FRAME))
+    tooltip:AddLine(Addon:SubjectDescription(Addon:Concat("+", L.SHIFT_KEY, L.RIGHT_CLICK), L.RESET_POSITION))
+    tooltip:AddLine(Addon:SubjectDescription(Addon:Concat("+", L.ALT_KEY, L.RIGHT_CLICK), Colors.Red(L.DESTROY_NEXT_ITEM)))
+    tooltip:Show()
+  end
 })
 
 Widgets:ConfigureForPointSync(frame, "MerchantButton")
@@ -37,6 +56,27 @@ frame:SetClickHandler("RightButton", "NONE", Commands.options)
 frame:SetClickHandler("RightButton", "SHIFT", function()
   StateManager:Dispatch(Actions:ResetMerchantButtonPoint())
 end)
+frame:SetClickHandler("RightButton", "ALT", Commands.destroy)
+
+-- Child frame.
+frame.childFrame = Widgets:Frame({
+  name = "$parent_ChildFrame",
+  parent = frame,
+  points = { { "CENTER" } }
+})
+frame.childFrame:SetBackdrop(nil)
+
+-- Icon.
+frame.icon = frame:CreateTexture("$parent_Icon", "ARTWORK")
+frame.icon:SetTexture(Addon:GetAsset("dejunk-icon"))
+frame.icon:SetPoint("LEFT", frame.childFrame)
+local iconSize = math.floor(frame:GetHeight()) - Widgets:Padding(0.5)
+frame.icon:SetSize(iconSize, iconSize)
+
+-- Reposition the label.
+frame.label:ClearAllPoints()
+frame.label:SetPoint("LEFT", frame.icon, "RIGHT", 0, 0)
+frame.label:SetJustifyH("LEFT")
 
 -- OnUpdate.
 frame:HookScript("OnUpdate", function(_, elapsed)
@@ -46,24 +86,13 @@ frame:HookScript("OnUpdate", function(_, elapsed)
   if frame.delayTimer < 0.1 then return end
   frame.delayTimer = 0
 
-  JunkFilter:GetSellableJunkItems(junkItems)
-  frame.label:SetText(LABEL_TEXT_FORMAT:format(#junkItems))
-end)
+  -- Update label text.
+  local numSellable, numDestroyable = JunkFilter:GetNumJunkItems()
+  frame.label:SetText(LABEL_TEXT_FORMAT:format(numSellable, numDestroyable))
 
--- OnEnter.
-frame:HookScript("OnEnter", function()
-  Tooltip:SetOwner(frame, "ANCHOR_RIGHT")
-  Tooltip:AddDoubleLine(Colors.Blue(ADDON_NAME), Colors.Grey(Addon.VERSION))
-  Tooltip:AddLine(Addon:SubjectDescription(L.LEFT_CLICK, L.START_SELLING))
-  Tooltip:AddLine(Addon:SubjectDescription(L.RIGHT_CLICK, L.TOGGLE_OPTIONS_FRAME))
-  Tooltip:AddLine(Addon:SubjectDescription(Addon:Concat("+", L.SHIFT_KEY, L.LEFT_CLICK), L.TOGGLE_JUNK_FRAME))
-  Tooltip:AddLine(Addon:SubjectDescription(Addon:Concat("+", L.SHIFT_KEY, L.RIGHT_CLICK), L.RESET_POSITION))
-  Tooltip:Show()
-end)
-
--- OnLeave.
-frame:HookScript("OnLeave", function()
-  Tooltip:Hide()
+  -- Resize child frame.
+  local width = frame.icon:GetWidth() + frame.label:GetStringWidth()
+  frame.childFrame:SetWidth(width)
 end)
 
 -- OnDragStart.